pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bayeslvr-ppo"
version = "0.1.0"
description = "BayesLVR-PPO for Uniswap v3 liquidity provision"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24,<3",
  "pandas>=2.0,<3",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
  "gymnasium>=0.29,<1.1",
  "stable-baselines3>=2.3,<3",
  "torch>=2.2",
  "ta-lib>=0.4.28",
  "shimmy>=1.3",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["src*"]
